Revert "Move utils that are only used in one plugin into their plugins."
This reverts commit 864e8e4981e6269ed203ed22ec5763561b8c1f80.
