+++ src/plugins/Dashboard/index.js
@@ -5,8 +5,8 @@ const DashboardUI = require('./Dashboard')
 const StatusBar = require('../StatusBar')
 const Informer = require('../Informer')
 const ThumbnailGenerator = require('../ThumbnailGenerator')
+const findAllDOMElements = require('../../utils/findAllDOMElements')
 const toArray = require('../../utils/toArray')
-const findAllDOMElements = require('./findAllDOMElements')
 const prettyBytes = require('prettier-bytes')
 const { defaultTabIcon } = require('./icons')
